Workaround fix for the call to self.scriptListTable.selectionModel().selectedIndexes() crashing Qt. All similar get-selected type methods crash the same way, but I had no trouble with any of the other methods.
Seems to be a lot of complaints about this on the web, and no definitive fix available.

Created a workaround which uses signals to tell the NetlistEditor class whenever the selected row in the GUI changes, and to just call removeRow() on whatever index NetlistEditor believes is selected, as opposed to asking Qt which one is selected (or, as it previously worked - getting Qt to crash).